Move cn to utils folder
